+++ src/automaticTB/properties/calc_mesh.py
@@ -97,3 +97,74 @@ def calculate_e_v_using_ibz(
 
     return full_e, full_v
 
+
+def calculate_e_c2_using_ibz(
+    tb: tightbinding.TightBindingModel, 
+    kpoints: np.ndarray, 
+    nks: typing.Tuple[int,int,int]
+) -> typing.Tuple[np.ndarray, np.ndarray]:
+    """calculates the kpoint energy using irreducible grid
+    
+    Note that the orbital component may not be simply the same for
+    equivalent k points. But should be the same if we do not distinguish
+    different atoms
+    """
+    cell = (tb.cell, tb.positions, tb.types)
+    sym_data = spglib.get_symmetry_dataset(cell, symprec=parameters.spgtol)
+
+    # point group operation
+    unique = []
+    for r in sym_data["rotations"]:
+        found = False
+
+        for r_u in unique:
+            if np.allclose(r, r_u, atol=parameters.stol): 
+                found = True
+                break
+        if found: continue
+        unique.append(r)
+    
+    c = tb.cell
+    rc = reciprocal.find_RCL(c)
+
+    nrot = len(unique)
+    rot_k_frac = np.zeros((nrot, 3, 3))
+    rot_k_cart = np.zeros((nrot, 3, 3))
+    for ir, r in enumerate(unique):
+        opc = c.T @ r @ np.linalg.inv(c.T)
+        rot_frac_k = np.linalg.inv(rc.T) @ opc @ rc.T
+        rot_k_cart[ir] = opc
+        rot_k_frac[ir] = rot_frac_k
+
+    nk = len(kpoints)
+    found_k = set()
+    symmetry_eq = np.zeros(nk, dtype=np.intp)
+    rotation_index = np.zeros(nk, dtype=np.intp)
+    
+    for i, k in enumerate(kpoints):
+        assert i == _get_k_index(k, nks)
+        if i in found_k: continue
+        rotated_k = rot_k_frac @ k
+        rotated_index = [_get_k_index(rk, nks) for rk in rotated_k]
+        for ir, rotated_i in enumerate(rotated_index):
+            if rotated_i in found_k: continue
+            rotation_index[rotated_i] = ir
+            symmetry_eq[rotated_i] = i
+            found_k.add(rotated_i)
+
+    unique_kid = np.unique(symmetry_eq)
+    unique_kpoints = kpoints[unique_kid]
+    rev = {u: iu for iu, u in enumerate(unique_kid)}
+
+    e, c = tb.solveE_at_ks(unique_kpoints)
+    c2 = np.abs(c)**2 # (nk, norb, nbnd)
+    _, nbnd = e.shape
+
+    full_e = np.zeros((nk, nbnd))
+    full_c2 = np.zeros((nk, nbnd, nbnd))
+    for i, (eqk, ri) in enumerate(zip(symmetry_eq, rotation_index)):
+        which_k = rev[eqk]
+        full_e[i, :] = e[which_k,:]
+        full_c2[i,:,:] = c2[which_k,:,:]
+
+    return full_e, full_c2